Initial version
This adds a new <ingredient> syntax, which can contains a standard list with
the list of ingredients, which can have multiple level.  Each line will be
parsed to try to understand it, and it can contain the following elements:
- an optional amount (float)
- an optional unit (supported: g, gr and ml), with an optional ".".  For now
  only supports those units as they are more or less equivalent, so we could in
  theory sum up everything with those unit and display a global weigth (however
  it's not done for now)
- a description for the ingredient

If all found unit are the same (g and gr being the same), an overall total will
be added to the rendered ingredient list.  Rows without unit are then assumed
to have the same unit.

If an ingredient has nested ingredients and the upper level has an amount, we
assume that it represent the overall amount for all nested ingredients and the
sum will just use it and ignore all amount underneath.

For instance, a list like:

<ingredients>
  * 20g of a
  * 50 g. of b
   * 10 of c
   * 40 gr. of d
  * 30 of e
</ingredients>

will display and overall "100g.".

It can also optionally handle variants, with a line starting with "option "
followed by any string naming that option.  If used, the first option must
appear before any ingredient.  All list will then be processed as independent
ingredient list, and a selector will be displayed to choose the wanted variant.

For instance:
<ingredients>
option Variant A
  * ...
option Variant B
  * ...
</ingredients>