Add ability to declare commands in ingredients markup.
The commands are executed once the ingredients lists are rendered, sequentially
and synchronously.  This is particularly useful when such an ingredient list is
generic and imported in other pages.  In that case you usually know you want
something specific, like "X grams of variant Y", so it's better to show that
directly.

The commands can appear anywhere in an ingredients markup and have one of those
general forms:

<ingredients>
[...]
cmd "what" X the command detail
cmd "what" the command detail
[...]
</ingredients>

Where "what" is the ingredient list name declared with an overall quantity
declaration ("pour" instruction), and where X is the optional sequence number
(1 based) to use in case multiple elements are found.  If not provided, default
is 1.

Note that element not visible (ie. their parent div as a "display: none" style)
are ignored in the numbering.

Two commands are available:
    - "variante X": to select the given X variant (using it's full text
      description)
    - "total X": to set the wanted weight for the autoamtic total weight

As the commands are executed sequentially and synchronously, you have to be
careful to first select a variant if any before changing its total.