Make amount optional in overall quantity declaration
This is useful if you want to be able to use a command on that ingredient list
even if its not for a specific quantity.

If a single overall quantity is declared and it doesn't have an amount, the
description is now used to better describe the automatic total weight.