Fix ingredient line parsing
If a unit is provided, with an optional "." or not, there should always be at
least one space after it before the rest of the line.

Not adding the space in the regexp group can make it ambiguous if the
description itself contains numbers / same caracters as unit when not itself
containing an amount.

Also make sure we parse the whole line, to avoid more ambiguity.