Have ING_CMD_SET_TOTAL reset the data-initval of all input fields.
The current behavior is for now safe as all ing_total (and therefore all
ingredient list) are independant.  So even if an ingredient list is included
from another page, modifying one won't touch the others and vice-versa.

But a future commit will introduce a feature to declare that a given full
recipe is meant for a certain mold dimensions and will allow to globally update
all ingredient lists in a page at the same time for a different mold dimension,
or set of molds, based on the volume ratios.  In that case, we want to take
into account the forced total of a given ingredient list, not the original one.